Thu, Dec 21, 2023, 4:35 PM -05:00

+++ half/half.c
@@ -18,7 +18,7 @@ int main(void)
 // TODO: Complete the function
     float half(float bill, float tax, int tip)
 {
-    long myHalf = (bill + (bill * tax * .01) + (bill * tip * .01)) / 2.0;
+    float myHalf = (bill + (bill * tax * .01) + (bill * tip * .01)) / 2.0;
     return myHalf;
 
 }
